Learning useContext & formatting scheduler
Made several changes in App.js and AppointmentPage.jsx all in hopes of implementing a cool useContext hook. Instead I was able to make an example of the hook work in the ServicePage --> ServiceCard hierarchy. Created a new file called "Calendar2.css" in order to modify the css of the react-calendar without changing the file in the node_modules due to the gitignore.